refactor: move load on scroll logic to custom hook
